refactor: move SelectedFilters in dedicated file
