+++ front/src/components/table/table-header/SortAndFilterBar.tsx
@@ -1,37 +1,13 @@
 import styled from '@emotion/styled';
-import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import SortOrFilterChip from './SortOrFilterChip';
 import { faArrowDown, faArrowUp } from '@fortawesome/pro-regular-svg-icons';
+import { SelectedFilterType, SelectedSortType } from './interface';
 
 type OwnProps<SortField> = {
   sorts: Array<SelectedSortType<SortField>>;
   onRemoveSort: (sortId: SelectedSortType<SortField>['id']) => void;
   filters: Array<SelectedFilterType>;
   onRemoveFilter: (filterId: SelectedFilterType['id']) => void;
-};
-
-export type SortType<SortIds = string> = {
-  label: string;
-  id: SortIds;
-  icon?: IconProp;
-};
-
-export type FilterType<KeyOfFilter = string> = {
-  label: string;
-  key: KeyOfFilter;
-  icon: IconProp;
-};
-
-export type SelectedFilterType = {
-  id: string;
-  label: string;
-  value: string;
-  operand: { id: string; label: string };
-  icon: IconProp;
-};
-
-export type SelectedSortType<SortField = string> = SortType<SortField> & {
-  order: 'asc' | 'desc';
 };
 
 const StyledBar = styled.div`
