+++ front/src/components/table/table-header/__stories__/SortDropdownButton.stories.tsx
@@ -1,4 +1,4 @@
-import { SelectedSortType, SortType } from '../SortAndFilterBar';
+import { SelectedSortType, SortType } from '../interface';
 import { ThemeProvider } from '@emotion/react';
 import { lightTheme } from '../../../../layout/styles/themes';
 import {
